feat(RevealCircleRepository.kt) : Location Repository의 이름 변경
